ai/requirements.txt
akshare==1.16.50
fastapi==0.115.11
openai==1.66.5
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn==0.34.0
yfinance==0.2.54
